2load_android_kernel: Add extra kernel cmdline from misc
This change allows for appending an extra command line string to the
kernel parameters based on the contents of the Android misc partition.

The `vb2ex_handle_android_misc_partition` function is modified to
populate a `vb2_android_misc_data` struct, which includes both the
boot mode and a dynamically composed command line string.

The `vb2_load_android` function is updated to use this struct. If an
extra command line is present, it is appended to the vendor command
line buffer. A buffer overflow check is included for safety.

The function's documentation and the stub implementation are also
updated to reflect these changes.

BRANCH=none
BUG=b:438666196
TEST=adb shell setprop arm64.memtag.bootctl memtag-once
     check /proc/cmdline in the next boot.

Cq-Depend: chromium:6979176
Change-Id: Ib8948dd6f9cd9ad6047176bffb1d87b4f6a4da10
Signed-off-by: Yidi Lin <yidilin@google.com>