2load_android_kernel: Allow adding extra kernel command line from misc partition
This change modifies the `vb2ex_handle_android_misc_partition` function
to allow it to return an additional command line string based on the
contents of the Android misc partition.

The `vb2_load_android` function now allocates a buffer, `extra_cmdline`,
for this extra command line. If  returned `extra_cmdline` is not empty,
it is appended to `params->vendor_cmdline_buffer`. A buffer overflow
check is included for safety.

The stub implementation is updated to handle the new parameters and
return an empty string.

BRANCH=none
BUG=b:438666196
TEST=adb shell setprop arm64.memtag.bootctl memtag-once
     check /proc/cmdline in the next boot.

Change-Id: Ib8948dd6f9cd9ad6047176bffb1d87b4f6a4da10
Signed-off-by: Yidi Lin <yidilin@google.com>