futility/updater: Preserve ME when falling back to full update
The 'preserve_me' quirk is supposed to be enabled for the auto-update
case only. However, in the edge case where we fall back from auto-update
to full update (when write protect is off and RO is different), the
quirk will still be incorrectly enabled. Fix that edge case via a new
field 'fallback_to_full_update' to record the fallback.

BUG=none
TEST=make runfutiltestscripts
BRANCH=none

Change-Id: I8688f9c9f23bc150b4562bede2db1e662193a2c5
Signed-off-by: Yu-Ping Wu <yupingso@chromium.org>