tests/futility/test_update: Make FROM_IMAGE and TO_IMAGE same board
Currently in most test cases, we update from a Peppy image to a Link
image. That is problematic, because their FMAP is different and
therefore incompatible. With b/462619916 we will add the FMAP
compatibility check to futility, so the test cases need to be fixed
first.

Here we change both FROM_IMAGE and TO_IMAGE to Trulo (Nissa) images. The
Nissa is selected for the following reasons:

* Nissa is an Intel device, which is required to run those ME-preserving
  test cases.
* Nissa has FMAP RO_GSCVD region, which is required to test the "AP RO
  locked" case.
* ifdtool supports locking/unlocking Nissa images.

The main differences between the new Trulo images and old Peppy and Link
images are:

* Unsigned Trulo images are used for easier modification, in case we
  need to modify and re-sign them for verification-related futility
  tests.
* Trulo images can be locked and unlocked by 'ifdtool -p adl'.

- unlock_me
- unlock eve me
- ifdtool
- archive
- custom label
- remove unused get_hwid because FROM_IMAGE and TO_IMAGE are the same.
- drop gbb12. v1.2 introduced in CL:221360 (landed in 6393.0.0)
- BOOT_STUB
- fix TO_IMAGE_WIPE_RW_VPD patch_file

Bash (and POSIX shells in general) are deeply integrated with C-style
string handling, where the null byte (\0) traditionally marks the end of
a string.

When the shell captures the output of the command substitution and
attempts to assign it as the argument $4 to the patch_file function, it
often treats the null byte as a terminator.

The single null byte is interpreted as an empty string.

So patch_file "$(printf '\000')" won't work.

BUG=none
TEST=make runfutiltestscripts
BRANCH=none

Change-Id: I4f9f5911207972314fdd1a3df5504c3764ec947c
Signed-off-by: Yu-Ping Wu <yupingso@chromium.org>