tests/futility/test_update: Make FROM_IMAGE and TO_IMAGE same board
- unlock_me
- unlock eve me
- ifdtool
- archive
- custom label
- remove unused get_hwid because FROM_IMAGE and TO_IMAGE are the same.
- drop gbb12. v1.2 introduced in CL:221360 (landed in 6393.0.0)
- fix TO_IMAGE_WIPE_RW_VPD patch_file

Bash (and POSIX shells in general) are deeply integrated with C-style
string handling, where the null byte (\0) traditionally marks the end of
a string.

When the shell captures the output of the command substitution and
attempts to assign it as the argument $4 to the patch_file function, it
often treats the null byte as a terminator.

The single null byte is interpreted as an empty string.

So patch_file "$(printf '\000')" won't work.

BUG=none
TEST=make runfutiltestscripts
BRANCH=none

Change-Id: I4f9f5911207972314fdd1a3df5504c3764ec947c
Signed-off-by: Yu-Ping Wu <yupingso@chromium.org>