crossystem: Fix crossystem lock file unlocking
crossystem was calling flock(<file>, F_UNLCK) to unlock a file. This is
not a correct behavior. Use LOCK_UN instead.
F_UNLCK = 3 and LOCK_UN = 8. Using F_UNLCK instead is equal to passing
a value ( LOCK_SH | LOCK_EX ). LibC implementations would treat it as
either LOCK_SH or LOCK_EX and the lock was usually used only in the
context of the same program instance. The OS was removing the lock after
the program exited. This what allowed this error to sit unnoticed for
a long time.
The issue was introduced in:
commit 58efaacd8ad2 ("crossystem: Add flock to prevent race conditions")

BRANCH=None
BUG=b:466079501
TEST=CQ
TEST=Read and set crossystem variables

Change-Id: Ie6ef3b4e9cf0516e4092f102b04fc8bef6a3fc82
Signed-off-by: Jakub Czapiga <czapiga@google.com>